调通ILruntime 下 Pb 的兼容，枚举和List<T>不要用，因为空List<T>无法反序列化，所以直接改为int和T[]传输 ILRuntime下lock尽量不要用，避免死锁！！！！所以XIHNET的kcpclient实现修改了一部分 ProtoBuf.PType.RegisterFunctionCreateInstance(PType_CreateInstance); ProtoBuf.PType.RegisterFunctionGetRealType(PType_GetRealType);         static object PType_CreateInstance(string typeName)         {             return Appdomain.Instantiate(typeName);         }
        static Type PType_GetRealType(object o)
        {
            var type = o.GetType();
            if (type.FullName == "ILRuntime.Runtime.Intepreter.ILTypeInstance")
            {
                var ilo = o as ILRuntime.Runtime.Intepreter.ILTypeInstance;
                type = ProtoBuf.PType.FindType(ilo.Type.FullName);
            }
            return type;
        }

------------------------------------------------------
static IMessageExt() {
            Type mt = typeof(IMessage);
#if XIHSERVER
            var types = mt.Assembly.GetTypes();
#else
            var types = XIHBasic.HotFixBridge.HotfixTypes;
#endif
            foreach (Type type in types)
            {
                if (type.IsAbstract || type.IsInterface || type.IsEnum || !type.IsSealed)
                {
                    continue;
                }
                if (!mt.IsAssignableFrom(type)) {//接口
                    continue;
                }
                object[] objects = type.GetCustomAttributes(typeof(MsgTypeCodeAttribute), false);
                if (objects.Length == 0)
                {
                    continue;
                }
                if (objects[0] is MsgTypeCodeAttribute attr) {
                    ushort val = attr.MsgTypeCode;
                    if (msg2Types.ContainsKey(val))
                    {
                        Debugger.Log($"<color=red>{msg2Types[val].FullName}与{type.FullName}类型的值一致【{val}】，将跳过此设置</color>");
                        continue;
                    }
                    msg2Types.Add(val, type);
                    type2Msgs.Add(type, val);
                    //Debugger.Log($"{val}:{type}");
#if !XIHSERVER
                    ProtoBuf.PType.RegisterType(type.FullName,type);
#endif
                    if (attr.IsResponse) {
                        rsps.Add(val);
                    }
                }
            }
        }
---------------------------------------------------------------------------

+++ XIHServer/XIHServer/Server/Handle/Login/LoginHandle.cs
@@ -26,7 +26,7 @@ namespace XIHServer
                 client.Send(new LoginRsp()
                 {
                     TaskId = req.TaskId,
-                    Result = LoginResultType.LoginResultError
+                    Result = (int)LoginResultType.LoginResultError
                 });
             }
             else
@@ -37,8 +37,8 @@ namespace XIHServer
                     TaskId = req.TaskId,
                     GateHost = Program.SvrCfg[NetServer.Lobby].IPEndPoint.Address.ToString(),
                     GatePort = Program.SvrCfg[NetServer.Lobby].IPEndPoint.Port,
-                    NetProtocol = Program.SvrCfg[NetServer.Lobby].NetProtocol,
-                    Result = LoginResultType.LoginResultSuccess,
+                    NetProtocol = (int)Program.SvrCfg[NetServer.Lobby].NetProtocol,
+                    Result = (int)LoginResultType.LoginResultSuccess,
                     SessionKey = sk
                 });
                 MockCache.WaitVerify.Add(sk, res.Value.Key);
